+++ common/src/main/java/com/chefmooon/ubesdelight/common/block/leaf_feast/base/SimpleLeafFeastBlock.java
@@ -1,6 +1,5 @@
 package com.chefmooon.ubesdelight.common.block.leaf_feast.base;
 
-import com.chefmooon.ubesdelight.UbesDelight;
 import com.chefmooon.ubesdelight.common.core.LeafFeastTypes;
 import com.chefmooon.ubesdelight.common.registry.UbesDelightBlocks;
 import com.chefmooon.ubesdelight.common.registry.UbesDelightDataComponentTypes;
@@ -69,8 +68,6 @@ public class SimpleLeafFeastBlock extends BaseLeafFeastBlock {
         } else {
             return tryAddItem(state, level, pos, player, hand);
         }
-
-//        return ItemInteractionResult.PASS_TO_DEFAULT_BLOCK_INTERACTION;
     }
 
     protected ItemInteractionResult tryAddItem(BlockState state, Level level, BlockPos pos, Player player, InteractionHand hand) {
@@ -88,6 +85,7 @@ public class SimpleLeafFeastBlock extends BaseLeafFeastBlock {
                     //spawnContainer(level, pos, state.getValue(FACING), heldItem); // todo - add and test when rice is added
                 }
                 playAddSound(level, pos);
+                LeafFeastBlock.triggerInsertAdvancement(player);
                 return ItemInteractionResult.SUCCESS;
             }
         }
@@ -115,8 +113,8 @@ public class SimpleLeafFeastBlock extends BaseLeafFeastBlock {
             level.setBlock(pos, state.setValue(SERVINGS, servings - 1), 3);
             playRemoveSound(level, pos);
             if (!player.isCreative()) {
-                if (player.isShiftKeyDown() && player.getFoodData().needsFood()) {
-                    player.eat(level, itemStack);
+                if (player.isShiftKeyDown() && (player.getFoodData().needsFood() || Objects.requireNonNull(itemStack.get(DataComponents.FOOD)).canAlwaysEat())) {
+                    tryEat(itemStack, level, pos, player);
                 } else {
                     if (!player.getInventory().add(itemStack)) {
                         player.drop(itemStack, false);
@@ -130,48 +128,8 @@ public class SimpleLeafFeastBlock extends BaseLeafFeastBlock {
             level.updateNeighbourForOutputSignal(pos, block);
             playRemoveSound(level, pos);
             if (!player.isCreative()) {
-                if (player.isShiftKeyDown() && player.getFoodData().needsFood()) {
-                    player.eat(level, itemStack);
-                } else {
-                    if (!player.getInventory().add(itemStack)) {
-                        player.drop(itemStack, false);
-                    }
-                }
-            }
-            return ItemInteractionResult.SUCCESS;
-        }
-
-        return ItemInteractionResult.FAIL;
-    }
-
-    protected ItemInteractionResult tryConsumeItem(BlockState state, Level level, BlockPos pos, Player player, InteractionHand hand) {
-        int servings = state.getValue(SERVINGS);
-        ItemStack itemStack = new ItemStack(servingItem.get());
-
-        UbesDelight.LOGGER.info("tryConsume: "  + player.isShiftKeyDown() + " | " + player.getFoodData().needsFood() + " | " + itemStack.has(DataComponents.FOOD));
-
-        if (servings > 1) {
-            level.setBlock(pos, state.setValue(SERVINGS, servings - 1), 3);
-            playRemoveSound(level, pos);
-            if (!player.isCreative()) {
-                if (player.isShiftKeyDown() && player.getFoodData().needsFood()) {
-                    player.eat(level, itemStack, Objects.requireNonNull(itemStack.get(DataComponents.FOOD)));
-                    UbesDelight.LOGGER.info("I should eat");
-                } else {
-                    if (!player.getInventory().add(itemStack)) {
-                        player.drop(itemStack, false);
-                    }
-                }
-            }
-            return ItemInteractionResult.SUCCESS;
-        } else if (servings == 1) {
-            Block block = BuiltInRegistryUtil.getBlock(UbesDelightBlocks.LEAF_FEAST);
-            level.setBlock(pos, getTransformState(block, state), 3);
-            level.updateNeighbourForOutputSignal(pos, block);
-            playRemoveSound(level, pos);
-            if (!player.isCreative()) {
-                if (player.isShiftKeyDown() && player.getFoodData().needsFood()) {
-                    player.eat(level, itemStack);
+                if (player.isShiftKeyDown() && (player.getFoodData().needsFood() || Objects.requireNonNull(itemStack.get(DataComponents.FOOD)).canAlwaysEat())) {
+                    tryEat(itemStack, level, pos, player);
                 } else {
                     if (!player.getInventory().add(itemStack)) {
                         player.drop(itemStack, false);