+++ common/src/main/java/com/chefmooon/ubesdelight/common/registry/UbesDelightAdvancements.java
@@ -1,7 +1,8 @@
 package com.chefmooon.ubesdelight.common.registry;
 
 import com.chefmooon.ubesdelight.common.advancement.BakingMatTrigger;
-import com.chefmooon.ubesdelight.common.advancement.LeafFeastTrigger;
+import com.chefmooon.ubesdelight.common.advancement.LeafFeastConsumeTrigger;
+import com.chefmooon.ubesdelight.common.advancement.LeafFeastInsertTrigger;
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import net.minecraft.advancements.CriterionTrigger;
 
@@ -9,7 +10,8 @@ import java.util.function.Supplier;
 
 public class UbesDelightAdvancements {
     public static final Supplier<BakingMatTrigger> USE_BAKING_MAT = registerTrigger("use_baking_mat", BakingMatTrigger::new);
-    public static final Supplier<LeafFeastTrigger> USE_LEAF_FEAST = registerTrigger("use_leaf_feast", LeafFeastTrigger::new);
+    public static final Supplier<LeafFeastInsertTrigger> USE_LEAF_FEAST_INSERT = registerTrigger("use_leaf_feast_insert", LeafFeastInsertTrigger::new);
+    public static final Supplier<LeafFeastConsumeTrigger> USE_LEAF_FEAST_CONSUME = registerTrigger("use_leaf_feast_consume",LeafFeastConsumeTrigger::new);
 
     @ExpectPlatform
     public static <T extends CriterionTrigger<?>> Supplier<T> registerTrigger(String name, Supplier<T> triggerSupplier) {