+++ neoforge/src/main/java/com/chefmooon/ubesdelight/common/block/leaf_feast/base/neoforge/UniversalLeafFeastBlockImpl.java
@@ -1,21 +1,27 @@
 package com.chefmooon.ubesdelight.common.block.leaf_feast.base.neoforge;
 
 import com.chefmooon.ubesdelight.common.block.entity.neoforge.UniversalLeafFeastBlockEntityImpl;
+import com.chefmooon.ubesdelight.common.block.leaf_feast.base.LeafFeastBlock;
 import com.chefmooon.ubesdelight.common.block.leaf_feast.base.UniversalLeafFeastBlock;
 import com.chefmooon.ubesdelight.common.core.LeafFeastTypes;
 import com.chefmooon.ubesdelight.common.registry.UbesDelightBlocks;
 import com.chefmooon.ubesdelight.common.utility.BuiltInRegistryUtil;
+import com.chefmooon.ubesdelight.common.utility.ItemStackUtil;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.world.Containers;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.ItemInteractionResult;
+import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
+
+import java.util.Objects;
 
 public class UniversalLeafFeastBlockImpl extends UniversalLeafFeastBlock {
     public UniversalLeafFeastBlockImpl(Properties properties) {
@@ -51,9 +57,10 @@ public class UniversalLeafFeastBlockImpl extends UniversalLeafFeastBlock {
             if (state.getValue(LEAF_FEAST_TYPE) == LeafFeastTypes.END || state.getValue(LEAF_FEAST_TYPE) == LeafFeastTypes.TIP) {
                 if (servings >= 3) return ItemInteractionResult.FAIL;
             }
-            if (universalLeafFeastBlockEntity.addItem(player.getAbilities().instabuild ? heldStack.copy() : heldStack)) {
+            if (universalLeafFeastBlockEntity.addItem(player, player.getAbilities().instabuild ? heldStack.copy() : heldStack)) {
                 level.setBlock(pos, state.setValue(SERVINGS, servings + 1), 3);
                 playAddSound(level, universalLeafFeastBlockEntity.getBlockPos());
+                LeafFeastBlock.triggerInsertAdvancement(player);
                 return ItemInteractionResult.SUCCESS;
             }
         }
@@ -71,10 +78,15 @@ public class UniversalLeafFeastBlockImpl extends UniversalLeafFeastBlock {
         }
 
         playRemoveSound(level, universalLeafFeastBlockEntity.getBlockPos());
-        if (player.isCreative()) {
-            universalLeafFeastBlockEntity.removeItem();
-        } else if (!player.getInventory().add(universalLeafFeastBlockEntity.removeItem())) {
-            Containers.dropItemStack(level, pos.getX(), pos.getY(), pos.getZ(), universalLeafFeastBlockEntity.removeItem());
+        ItemStack itemStack = universalLeafFeastBlockEntity.removeItem();
+        if (!player.isCreative()) {
+            if (player.isShiftKeyDown() && (player.getFoodData().needsFood() || Objects.requireNonNull(itemStack.get(DataComponents.FOOD)).canAlwaysEat())) {
+                tryEat(itemStack, level, pos, player);
+            } else {
+                if (!player.getInventory().add(itemStack)) {
+                    Containers.dropItemStack(level, pos.getX(), pos.getY(), pos.getZ(), itemStack);
+                }
+            }
         }
 
         if (servings == 1) {
@@ -84,6 +96,24 @@ public class UniversalLeafFeastBlockImpl extends UniversalLeafFeastBlock {
         }
 
         return ItemInteractionResult.SUCCESS;
+    }
+
+    public static void tryEat(ItemStack itemStack, Level level, BlockPos pos, Player player) {
+        ItemStack container = ItemStackUtil.getContainer(itemStack);
+        if (!container.isEmpty()) {
+            spawnContainer(level, pos, player.getDirection().getOpposite(), container);
+        }
+        player.eat(level, itemStack);
+        LeafFeastBlock.triggerConsumeAdvancement(player);
+    }
+
+    public static void spawnContainer(Level level, BlockPos blockPos, Direction direction, ItemStack itemStack) {
+        ItemEntity entity = new ItemEntity(level,
+                blockPos.getX() + 0.5 + (direction.getStepX() * 0.2),
+                blockPos.getY() + 0.2,
+                blockPos.getZ() + 0.5 + (direction.getStepZ() * 0.2), itemStack.copy());
+        entity.setDeltaMovement(direction.getStepX() * 0.2F, 0.0F, direction.getStepZ() * 0.2F);
+        level.addFreshEntity(entity);
     }
 
     @Override